Created form response schema & model
Created a form response model that is based on this schema:

-- questionAnswerSchema --

answer:	Array<String>

-- responseSchema --

respondeeEmail:	String
formId:		String
answers:	Map<String, questionAnswerSchema>

In questionAnswerSchema `answer` is an array of strings for holding
multiple answers of the same question, through checkboxes.

Eventhough a question is text-based, it would still be stored in the
array.

An answer of a text question would be:
answer: ['I like cats because they reduce my stress and anxiety']

So access it with index 0.

Whereas an answer of a checkbox question would be:
answer: ['Cats', 'Dogs', 'Rabbits', 'Birds']